replace example with ckn-oracle in integration tests
