--- examples/tests/requirements.txt
@@ -1,2 +1,0 @@
-neo4j
-pytest
